rename namespace "ConfigGUI" to "ConfigurationAttribute"
