+++ Sync/Sync.csproj
@@ -98,7 +98,7 @@
     <Compile Include="MessageFilter\MessageDispatcher.cs" />
     <Compile Include="MessageFilter\MessageManager.cs" />
     <Compile Include="Tools\Builtin\CommonCommand.cs" />
-    <Compile Include="Tools\ConfigGUI\ClientAndSourceAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\ClientAndSourceAttribute.cs" />
     <Compile Include="Tools\Builtin\PluginCommand.cs" />
     <Compile Include="Tools\Builtin\InternalPlugin.cs" />
     <Compile Include="Tools\SyncIO\FileLoggerWriter.cs" />
@@ -114,16 +114,16 @@
     <Compile Include="Source\SourceBase.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Command\CommandDispatch.cs" />
-    <Compile Include="Tools\ConfigGUI\BoolAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\ColorAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\ConfigAttributeBase.cs" />
-    <Compile Include="Tools\ConfigGUI\GuiLanguageElement.cs" />
-    <Compile Include="Tools\ConfigGUI\FloatAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\FontAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\IntegerAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\ListAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\PathAttribute.cs" />
-    <Compile Include="Tools\ConfigGUI\StringAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\BoolAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\ColorAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\ConfigAttributeBase.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\GuiLanguageElement.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\FloatAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\FontAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\IntegerAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\ListAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\PathAttribute.cs" />
+    <Compile Include="Tools\ConfigurationAttribute\StringAttribute.cs" />
     <Compile Include="Tools\Configuration.cs" />
     <Compile Include="Tools\ConfigurationIO.cs" />
     <Compile Include="Tools\SyncIO\ConsoleWriter.cs" />
