+++ Sync/Tools/ConfigurationAttribute/GuiLanguageElement.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Sync.Tools.ConfigGUI
+namespace Sync.Tools.ConfigurationAttribute
 {
     public struct GuiLanguageElement
     {
