+++ Sync/Tools/I18n.cs
@@ -1,4 +1,4 @@
-﻿using Sync.Tools.ConfigGUI;
+﻿using Sync.Tools.ConfigurationAttribute;
 using System;
 using System.Collections.Generic;
 using System.IO;
